Adding PlayerMove class, adding another player constructor with a location field
I haven't tested my playerMove class much, this is just an early draft
and should not be included in any final product until tested. Only the
actual movement actions have been implimented.

Basically this upload is just to share what I have so far and let all
of you point out any glaring errors :)

Tomorrow I'll be testing each of the methods more thoroughly as well as
implimenting the rest of the undefined methods in the PlayerMove class.

The last thing I want to do is write a bunch of code we can't use.