Adding cureDisease to PlayerMove class to work with cureStatusIdeas.cpp
I added a cureDisease method to PlayerMove class that will check
conditions and update the player hand after a discard. I made it with
the thought of calling the cureDisease function in cureStatusIdeas.cpp.
They do both have the same name… which is a bit sketchy. Maybe we could
call the other function "updateCureStatus(string)" or something of that
nature?