Modified Deck Class to prevent EPIDEMIC cards from being in the first ten cards of the deck.
        int epics[EPICS] = {};
        for ( size_t i = 0; i < EPICS; i++ ) {
            size_t r = DEALT + ( ( ceCards - DEALT ) / EPICS * i ) +
                     rand() % ( ( ceCards - DEALT ) / EPICS );
            epics[i] = r;
        }

DEALT is a constant set to "9" in Deck.h and represents the cards
designated to the players, so this routine places (currently) six
EPIDEMICS in the following possible index ranges of the deck:

    1 - 10..16
    2 - 16..22
    3 - 23..29
    4 - 30..36
    5 - 37..43
    6 - 44..50

So, an EPIDEMIC is certain to not be at the bottom of the deck,
but it could very well be the first card dealt in a three player
game.  This shuld be sufficient.  Now, I need to adjust the
number of total EPIDEMIC cards in deck to relate to the difficulty
level chosen at the start of the game.

--Derrick